+++ src/buttons/flat-button-label.jsx
@@ -1,11 +1,14 @@
 const React = require('react/addons');
-const PureRenderMixin = React.addons.PureRenderMixin;
+const ContextPure = require('../mixins/context-pure');
 const Styles = require('../utils/styles');
-
+const DefaultRawTheme = require('../styles/raw-themes/light-raw-theme');
+const ThemeManager = require('../styles/theme-manager');
 
 const FlatButtonLabel = React.createClass({
 
-  mixins: [PureRenderMixin],
+  mixins: [
+    ContextPure,
+  ],
 
   contextTypes: {
     muiTheme: React.PropTypes.object,
@@ -16,12 +19,36 @@ const FlatButtonLabel = React.createClass({
     style: React.PropTypes.object,
   },
 
-  getContextProps() {
-    const theme = this.context.muiTheme;
+  //for passing default theme context to children
+  childContextTypes: {
+    muiTheme: React.PropTypes.object,
+  },
 
+  getChildContext () {
     return {
-      spacingDesktopGutterLess: theme.spacing.desktopGutterLess,
+      muiTheme: this.state.muiTheme,
     };
+  },
+
+  getInitialState () {
+    return {
+      muiTheme: this.context.muiTheme ? this.context.muiTheme : ThemeManager.getMuiTheme(DefaultRawTheme),
+    };
+  },
+
+  //to update theme inside state whenever a new theme is passed down
+  //from the parent / owner using context
+  componentWillReceiveProps (nextProps, nextContext) {
+    let newMuiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
+    this.setState({muiTheme: newMuiTheme});
+  },
+
+  statics: {
+    getRelevantContextKeys(muiTheme) {
+      return {
+        spacingDesktopGutterLess: muiTheme.rawTheme.spacing.desktopGutterLess,
+      };
+    },
   },
 
   render: function() {
@@ -30,11 +57,11 @@ const FlatButtonLabel = React.createClass({
       style,
     } = this.props;
 
-    const contextProps = this.getContextProps();
+    const contextKeys = this.constructor.getRelevantContextKeys(this.state.muiTheme);
 
     const mergedRootStyles = Styles.mergeAndPrefix({
       position: 'relative',
-      padding: '0 ' + contextProps.spacingDesktopGutterLess + 'px',
+      padding: '0 ' + contextKeys.spacingDesktopGutterLess + 'px',
     }, style);
 
     return (