+++ src/card/card-actions.jsx
@@ -1,8 +1,8 @@
-let React = require('react');
-let StylePropable = require('../mixins/style-propable');
+const React = require('react');
+const StylePropable = require('../mixins/style-propable');
 
 
-let CardActions = React.createClass({
+const CardActions = React.createClass({
   mixins: [StylePropable],
 
   getStyles() {