+++ src/date-picker/calendar-month.jsx
@@ -1,10 +1,10 @@
-let React = require('react');
-let DateTime = require('../utils/date-time');
-let DayButton = require('./day-button');
-let ClearFix = require('../clearfix');
+const React = require('react');
+const DateTime = require('../utils/date-time');
+const DayButton = require('./day-button');
+const ClearFix = require('../clearfix');
 
 
-let CalendarMonth = React.createClass({
+const CalendarMonth = React.createClass({
 
   propTypes: {
     displayDate: React.PropTypes.object.isRequired,