+++ src/flat-button.jsx
@@ -1,5 +1,5 @@
 const React = require('react/addons');
-const PureRenderMixin = React.addons.PureRenderMixin;
+const ContextPure = require('./mixins/context-pure');
 const Transitions = require('./styles/transitions');
 const Children = require('./utils/children');
 const ColorManipulator = require('./utils/color-manipulator');
@@ -7,7 +7,8 @@ const ImmutabilityHelper = require('./utils/immutability-helper');
 const Typography = require('./styles/typography');
 const EnhancedButton = require('./enhanced-button');
 const FlatButtonLabel = require('./buttons/flat-button-label');
-
+const DefaultRawTheme = require('./styles/raw-themes/light-raw-theme');
+const ThemeManager = require('./styles/theme-manager');
 
 function validateLabel (props, propName, componentName) {
   if (!props.children && !props.label) {
@@ -18,10 +19,48 @@ function validateLabel (props, propName, componentName) {
 
 const FlatButton = React.createClass({
 
-  mixins: [PureRenderMixin],
+  mixins: [
+    ContextPure,
+  ],
+
+  statics: {
+    getRelevantContextKeys(muiTheme) {
+      const buttonTheme = muiTheme.button;
+      const flatButtonTheme = muiTheme.flatButton;
+
+      return {
+        buttonColor: flatButtonTheme.color,
+        buttonHeight: buttonTheme.height,
+        buttonMinWidth: buttonTheme.minWidth,
+        disabledTextColor: flatButtonTheme.disabledTextColor,
+        primaryTextColor: flatButtonTheme.primaryTextColor,
+        secondaryTextColor: flatButtonTheme.secondaryTextColor,
+        textColor: flatButtonTheme.textColor,
+        textTransform: flatButtonTheme.textTransform ? flatButtonTheme.textTransform :
+                      (buttonTheme.textTransform ? buttonTheme.textTransform : 'uppercase'),
+      };
+    },
+    getChildrenClasses() {
+      return [
+        EnhancedButton,
+        FlatButtonLabel,
+      ];
+    },
+  },
 
   contextTypes: {
     muiTheme: React.PropTypes.object,
+  },
+
+  //for passing default theme context to children
+  childContextTypes: {
+    muiTheme: React.PropTypes.object,
+  },
+
+  getChildContext () {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
   },
 
   propTypes: {
@@ -55,23 +94,15 @@ const FlatButton = React.createClass({
       hovered: false,
       isKeyboardFocused: false,
       touch: false,
+      muiTheme: this.context.muiTheme ? this.context.muiTheme : ThemeManager.getMuiTheme(DefaultRawTheme),
     };
   },
 
-  getContextProps() {
-    const theme = this.context.muiTheme;
-    const buttonTheme = theme.component.button;
-    const flatButtonTheme = theme.component.flatButton;
-
-    return {
-      buttonColor: flatButtonTheme.color,
-      buttonHeight: buttonTheme.height,
-      buttonMinWidth: buttonTheme.minWidth,
-      disabledTextColor: flatButtonTheme.disabledTextColor,
-      primaryTextColor: flatButtonTheme.primaryTextColor,
-      secondaryTextColor: flatButtonTheme.secondaryTextColor,
-      textColor: flatButtonTheme.textColor,
-    };
+  //to update theme inside state whenever a new theme is passed down
+  //from the parent / owner using context
+  componentWillReceiveProps (nextProps, nextContext) {
+    let newMuiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
+    this.setState({muiTheme: newMuiTheme});
   },
 
   render() {
@@ -79,6 +110,7 @@ const FlatButton = React.createClass({
       children,
       disabled,
       hoverColor,
+      backgroundColor,
       label,
       labelStyle,
       labelPosition,
@@ -93,33 +125,43 @@ const FlatButton = React.createClass({
       ...other,
       } = this.props;
 
-    const contextProps = this.getContextProps();
+    const {
+      buttonColor,
+      buttonHeight,
+      buttonMinWidth,
+      disabledTextColor,
+      primaryTextColor,
+      secondaryTextColor,
+      textColor,
+      textTransform,
+    } = this.constructor.getRelevantContextKeys(this.state.muiTheme);
 
-    const defaultColor = disabled ? contextProps.disabledTextColor :
-      primary ? contextProps.primaryTextColor :
-      secondary ? contextProps.secondaryTextColor :
-      contextProps.textColor;
+    const defaultColor = disabled ? disabledTextColor :
+      primary ? primaryTextColor :
+      secondary ? secondaryTextColor :
+      textColor;
 
     const defaultHoverColor = ColorManipulator.fade(ColorManipulator.lighten(defaultColor, 0.4), 0.15);
     const defaultRippleColor = ColorManipulator.fade(defaultColor, 0.8);
     const buttonHoverColor = hoverColor || defaultHoverColor;
     const buttonRippleColor = rippleColor || defaultRippleColor;
     const hovered = (this.state.hovered || this.state.isKeyboardFocused) && !disabled;
+    const buttonBackgroundColor = backgroundColor || buttonColor;
 
     const mergedRootStyles = ImmutabilityHelper.merge({
       color: defaultColor,
       transition: Transitions.easeOut(),
       fontSize: Typography.fontStyleButtonFontSize,
       letterSpacing: 0,
-      textTransform: 'uppercase',
+      textTransform: textTransform,
       fontWeight: Typography.fontWeightMedium,
       borderRadius: 2,
       userSelect: 'none',
       position: 'relative',
       overflow: 'hidden',
-      backgroundColor: hovered ? buttonHoverColor : contextProps.buttonColor,
-      lineHeight: contextProps.buttonHeight + 'px',
-      minWidth: contextProps.buttonMinWidth,
+      backgroundColor: hovered ? buttonHoverColor : buttonBackgroundColor,
+      lineHeight: buttonHeight + 'px',
+      minWidth: buttonMinWidth,
       padding: 0,
       margin: 0,
       //This is need so that ripples do not bleed past border radius.