+++ src/menu/menu-item.jsx
@@ -1,7 +1,9 @@
-let React = require('react');
-let StylePropable = require('../mixins/style-propable');
-let FontIcon = require('../font-icon');
-let Toggle = require('../toggle');
+const React = require('react');
+const StylePropable = require('../mixins/style-propable');
+const FontIcon = require('../font-icon');
+const Toggle = require('../toggle');
+const DefaultRawTheme = require('../styles/raw-themes/light-raw-theme');
+const ThemeManager = require('../styles/theme-manager');
 
 const Types = {
   LINK: 'LINK',
@@ -10,7 +12,7 @@ const Types = {
 };
 
 
-let MenuItem = React.createClass({
+const MenuItem = React.createClass({
 
   mixins: [StylePropable],
 
@@ -36,6 +38,30 @@ let MenuItem = React.createClass({
     active: React.PropTypes.bool,
   },
 
+  //for passing default theme context to children
+  childContextTypes: {
+    muiTheme: React.PropTypes.object,
+  },
+
+  getChildContext () {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
+  getInitialState () {
+    return {
+      muiTheme: this.context.muiTheme ? this.context.muiTheme : ThemeManager.getMuiTheme(DefaultRawTheme),
+    };
+  },
+
+  //to update theme inside state whenever a new theme is passed down
+  //from the parent / owner using context
+  componentWillReceiveProps (nextProps, nextContext) {
+    let newMuiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
+    this.setState({muiTheme: newMuiTheme});
+  },
+
   statics: {
     Types: Types,
   },
@@ -49,11 +75,11 @@ let MenuItem = React.createClass({
   },
 
   getTheme() {
-    return this.context.muiTheme.component.menuItem;
+    return this.state.muiTheme.menuItem;
   },
 
   getSpacing() {
-    return this.context.muiTheme.spacing;
+    return this.state.muiTheme.rawTheme.spacing;
   },
 
   getStyles() {
@@ -64,7 +90,7 @@ let MenuItem = React.createClass({
         lineHeight: this.getTheme().height + 'px',
         paddingLeft: this.getTheme().padding,
         paddingRight: this.getTheme().padding,
-        color: this.context.muiTheme.palette.textColor,
+        color: this.state.muiTheme.rawTheme.palette.textColor,
       },
       number: {
         float: 'right',
@@ -92,10 +118,10 @@ let MenuItem = React.createClass({
         top: -12,
         position: 'relative',
         fontWeight: 300,
-        color: this.context.muiTheme.palette.textColor,
+        color: this.state.muiTheme.rawTheme.palette.textColor,
       },
       toggle: {
-        marginTop: ((this.getTheme().height - this.context.muiTheme.component.radioButton.size) / 2),
+        marginTop: ((this.getTheme().height - this.state.muiTheme.radioButton.size) / 2),
         float: 'right',
         width: 42,
       },
@@ -107,7 +133,7 @@ let MenuItem = React.createClass({
       },
       rootWhenDisabled: {
         cursor: 'default',
-        color: this.context.muiTheme.palette.disabledColor,
+        color: this.state.muiTheme.rawTheme.palette.disabledColor,
       },
     };
 
@@ -161,9 +187,9 @@ let MenuItem = React.createClass({
 
         {icon}
         {this.props.children}
-        {data}
-        {attribute}
         {number}
+        {attribute}
+        {data}
         {toggleElement}
         {iconRight}