+++ src/text-field.jsx
@@ -1,9 +1,12 @@
-let React = require('react');
-let ColorManipulator = require('./utils/color-manipulator');
-let StylePropable = require('./mixins/style-propable');
-let Transitions = require('./styles/transitions');
-let UniqueId = require('./utils/unique-id');
-let EnhancedTextarea = require('./enhanced-textarea');
+const React = require('react');
+const ColorManipulator = require('./utils/color-manipulator');
+const StylePropable = require('./mixins/style-propable');
+const Transitions = require('./styles/transitions');
+const UniqueId = require('./utils/unique-id');
+const EnhancedTextarea = require('./enhanced-textarea');
+const DefaultRawTheme = require('./styles/raw-themes/light-raw-theme');
+const ThemeManager = require('./styles/theme-manager');
+const ContextPure = require('./mixins/context-pure');
 
 /**
  * Check if a value is valid to be displayed inside an input.
@@ -15,9 +18,12 @@ function isValid(value) {
   return value || value === 0;
 }
 
-let TextField = React.createClass({
+const TextField = React.createClass({
 
-  mixins: [StylePropable],
+  mixins: [
+    ContextPure,
+    StylePropable,
+  ],
 
   contextTypes: {
     muiTheme: React.PropTypes.object,
@@ -33,6 +39,7 @@ let TextField = React.createClass({
       React.PropTypes.string,
       React.PropTypes.element,
     ]),
+    hintStyle: React.PropTypes.object,
     id: React.PropTypes.string,
     inputStyle: React.PropTypes.object,
     multiLine: React.PropTypes.bool,
@@ -42,10 +49,22 @@ let TextField = React.createClass({
     onFocus: React.PropTypes.func,
     onKeyDown: React.PropTypes.func,
     rows: React.PropTypes.number,
+    rowsMax: React.PropTypes.number,
     type: React.PropTypes.string,
     underlineStyle: React.PropTypes.object,
     underlineFocusStyle: React.PropTypes.object,
     underlineDisabledStyle: React.PropTypes.object,
+  },
+
+  //for passing default theme context to children
+  childContextTypes: {
+    muiTheme: React.PropTypes.object,
+  },
+
+  getChildContext () {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
   },
 
   getDefaultProps() {
@@ -56,12 +75,27 @@ let TextField = React.createClass({
     };
   },
 
-  getContextProps() {
-    const theme = this.context.muiTheme;
+  statics: {
+    getRelevantContextKeys(muiTheme) {
+      const textFieldTheme = muiTheme.textField
 
-    return {
-      isRtl: theme.isRtl,
-    };
+      return {
+        floatingLabelColor: textFieldTheme.floatingLabelColor,
+        focusColor: textFieldTheme.focusColor,
+        borderColor: textFieldTheme.borderColor,
+        textColor: textFieldTheme.textColor,
+        disabledTextColor: textFieldTheme.disabledTextColor,
+        backgroundColor: textFieldTheme.backgroundColor,
+        hintColor: textFieldTheme.hintColor,
+        errorColor: textFieldTheme.errorColor,
+        isRtl: muiTheme.isRtl,
+      };
+    },
+    getChildrenClasses() {
+      return [
+        EnhancedTextarea,
+      ];
+    },
   },
 
   getInitialState() {
@@ -71,19 +105,17 @@ let TextField = React.createClass({
       errorText: this.props.errorText,
       hasValue: isValid(props.value) || isValid(props.defaultValue) ||
         (props.valueLink && isValid(props.valueLink.value)),
+      muiTheme: this.context.muiTheme ? this.context.muiTheme : ThemeManager.getMuiTheme(DefaultRawTheme),
     };
-  },
-
-  getTheme() {
-    return this.context.muiTheme.component.textField;
   },
 
   componentDidMount() {
     this._uniqueId = UniqueId.generate();
   },
 
-  componentWillReceiveProps(nextProps) {
+  componentWillReceiveProps(nextProps, nextContext) {
     let newState = {};
+    newState.muiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
 
     newState.errorText = nextProps.errorText;
     if (nextProps.children && nextProps.children.props) {
@@ -108,9 +140,18 @@ let TextField = React.createClass({
   },
 
   getStyles() {
-    let props = this.props;
-    let theme = this.getTheme();
-    const contextProps = this.getContextProps();
+    const props = this.props;
+    const {
+      floatingLabelColor,
+      focusColor,
+      borderColor,
+      textColor,
+      disabledTextColor,
+      backgroundColor,
+      hintColor,
+      errorColor,
+      isRtl,
+    } = this.constructor.getRelevantContextKeys(this.state.muiTheme);
 
     let styles = {
       root: {
@@ -120,7 +161,7 @@ let TextField = React.createClass({
         height: (props.rows - 1) * 24 + (props.floatingLabelText ? 72 : 48),
         display: 'inline-block',
         position: 'relative',
-        fontFamily: this.context.muiTheme.contentFontFamily,
+        fontFamily: this.state.muiTheme.rawTheme.fontFamily,
         transition: Transitions.easeOut('200ms', 'height'),
       },
       error: {
@@ -128,14 +169,14 @@ let TextField = React.createClass({
         bottom: 5,
         fontSize: 12,
         lineHeight: '12px',
-        color: theme.errorColor,
+        color: errorColor,
         transition: Transitions.easeOut(),
       },
       hint: {
         position: 'absolute',
         lineHeight: '22px',
         opacity: 1,
-        color: theme.hintColor,
+        color: hintColor,
         transition: Transitions.easeOut(),
         bottom: 12,
       },
@@ -147,13 +188,13 @@ let TextField = React.createClass({
         height: '100%',
         border: 'none',
         outline: 'none',
-        backgroundColor: theme.backgroundColor,
-        color: props.disabled ? theme.disabledTextColor : theme.textColor,
+        backgroundColor: backgroundColor,
+        color: props.disabled ? disabledTextColor : textColor,
         font: 'inherit',
       },
       underline: {
         border: 'none',
-        borderBottom: 'solid 1px ' + theme.borderColor,
+        borderBottom: 'solid 1px ' + borderColor,
         position: 'absolute',
         width: '100%',
         bottom: 8,
@@ -169,11 +210,11 @@ let TextField = React.createClass({
         userSelect: 'none',
         cursor: 'default',
         bottom: 8,
-        borderBottom: 'dotted 2px ' + theme.disabledTextColor,
+        borderBottom: 'dotted 2px ' + disabledTextColor,
       },
       underlineFocus: {
         borderBottom: 'solid 2px',
-        borderColor: theme.focusColor,
+        borderColor: focusColor,
         transform: 'scaleX(0)',
         transition: Transitions.easeOut(),
       },
@@ -189,7 +230,7 @@ let TextField = React.createClass({
       bottom: 'none',
       opacity: 1,
       transform: 'scale(1) translate3d(0, 0, 0)',
-      transformOrigin: contextProps.isRtl ? 'right top' : 'left top',
+      transformOrigin: isRtl ? 'right top' : 'left top',
     });
 
     styles.textarea = this.mergeStyles(styles.input, {
@@ -202,13 +243,13 @@ let TextField = React.createClass({
     styles.focusUnderline = this.mergeStyles(styles.underline, styles.underlineFocus, props.underlineFocusStyle);
 
     if (this.state.isFocused) {
-      styles.floatingLabel.color = theme.focusColor;
+      styles.floatingLabel.color = focusColor;
       styles.floatingLabel.transform = 'perspective(1px) scale(0.75) translate3d(2px, -28px, 0)';
       styles.focusUnderline.transform = 'scaleX(1)';
     }
 
     if (this.state.hasValue) {
-      styles.floatingLabel.color = ColorManipulator.fade(props.disabled ? theme.disabledTextColor : theme.floatingLabelColor, 0.5);
+      styles.floatingLabel.color = ColorManipulator.fade(props.disabled ? disabledTextColor : floatingLabelColor, 0.5);
       styles.floatingLabel.transform = 'perspective(1px) scale(0.75) translate3d(2px, -28px, 0)';
       styles.hint.opacity = 0;
     }
@@ -224,7 +265,7 @@ let TextField = React.createClass({
     }
 
     if (this.state.errorText && this.state.isFocused) styles.floatingLabel.color = styles.error.color;
-    if (props.floatingLabelText && !props.multiLine) styles.input.paddingTop = 26;
+    if (props.floatingLabelText && !props.multiLine) styles.input.marginTop = 14;
 
     if (this.state.errorText) {
       styles.focusUnderline.borderColor = styles.error.color;
@@ -242,6 +283,7 @@ let TextField = React.createClass({
       floatingLabelText,
       fullWidth,
       hintText,
+      hintStyle,
       id,
       multiLine,
       onBlur,
@@ -249,6 +291,7 @@ let TextField = React.createClass({
       onFocus,
       type,
       rows,
+      rowsMax,
       ...other,
     } = this.props;
 
@@ -261,7 +304,7 @@ let TextField = React.createClass({
     ) : null;
 
     let hintTextElement = hintText ? (
-      <div style={this.mergeAndPrefix(styles.hint)}>{hintText}</div>
+      <div style={this.mergeAndPrefix(styles.hint, hintStyle)}>{hintText}</div>
     ) : null;
 
     let floatingLabelTextElement = floatingLabelText ? (
@@ -296,7 +339,8 @@ let TextField = React.createClass({
         <EnhancedTextarea
           {...other}
           {...inputProps}
-        rows={rows}
+          rows={rows}
+          rowsMax={rowsMax}
           onHeightChange={this._handleTextAreaHeightChange}
           textareaStyle={this.mergeAndPrefix(styles.textarea)} />
       ) : (