+++ src/toolbar/toolbar-title.jsx
@@ -1,8 +1,9 @@
-let React = require('react');
-let StylePropable = require('../mixins/style-propable');
+const React = require('react');
+const StylePropable = require('../mixins/style-propable');
+const DefaultRawTheme = require('../styles/raw-themes/light-raw-theme');
+const ThemeManager = require('../styles/theme-manager');
 
-
-let ToolbarTitle = React.createClass({
+const ToolbarTitle = React.createClass({
 
   mixins: [StylePropable],
 
@@ -14,8 +15,32 @@ let ToolbarTitle = React.createClass({
     text: React.PropTypes.string,
   },
 
+  //for passing default theme context to children
+  childContextTypes: {
+    muiTheme: React.PropTypes.object,
+  },
+
+  getChildContext () {
+    return {
+      muiTheme: this.state.muiTheme,
+    };
+  },
+
+  getInitialState () {
+    return {
+      muiTheme: this.context.muiTheme ? this.context.muiTheme : ThemeManager.getMuiTheme(DefaultRawTheme),
+    };
+  },
+
+  //to update theme inside state whenever a new theme is passed down
+  //from the parent / owner using context
+  componentWillReceiveProps (nextProps, nextContext) {
+    let newMuiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
+    this.setState({muiTheme: newMuiTheme});
+  },
+
   getTheme() {
-    return this.context.muiTheme.component.toolbar;
+    return this.state.muiTheme.toolbar;
   },
 
   render() {
@@ -26,7 +51,7 @@ let ToolbarTitle = React.createClass({
     } = this.props;
 
     let styles = this.mergeAndPrefix({
-      paddingRight: this.context.muiTheme.spacing.desktopGutterLess,
+      paddingRight: this.state.muiTheme.rawTheme.spacing.desktopGutterLess,
       lineHeight: this.getTheme().height + 'px',
       fontSize: this.getTheme().titleFontSize + 'px',
       display: 'inline-block',